back/pyproject.toml
[tool.poetry]
name = "speech-gpt-web-back"
version = "0.1.0"
description = "Speech-GPT Web Application Backend"
authors = ["reflaxess123 <reflaxess@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.14"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
websockets = "^14.1"
httpx = "^0.28.1"
python-multipart = "^0.0.20"
pydantic = "^2.10.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
